requirements.txt
groq>=0.4.0
requests>=2.28.0
python-docx>=0.8.11
ddgs>=6.0.0
beautifulsoup4>=4.12.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.0.0
pyttsx3>=2.90
pywin32>=306
psutil>=5.9.0
schedule>=1.2.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.14
